Ajout de long pour les id

+++ src/M3 Pratique/Etat.cs
@@ -10,10 +10,10 @@ namespace M3_Pratique
     // Classe représentatnt un état
     public class Etat
     {
-        private int _id;
+        private long _id;
         private string _libelle;
 
-        public int Id { get => _id; set => _id = value; }
+        public long Id { get => _id; set => _id = value; }
         public string Libelle { get => _libelle; set => _libelle = value; }
 
         // Constructeur d'un état
